refactor: #2097 extract activity-log-aggregation to resolve service import cycle
PR #2124 (ADR-0048 week 4) で Tier 2 Review BLOCK [Fix 2] 対応。
biome noImportCycles 2 件 fail を解消する。

原因: activity-log-service と auto-challenge-service が相互依存
  - activity-log-service.recordActivity() → auto-challenge-service.incrementChallengeProgress()
    (記録 → チャレンジ進捗更新)
  - auto-challenge-service.analyzeWeakCategory() → activity-log-service.getActivityLogs()
    (チャレンジ生成 → 過去ログ集計)
両方向の dynamic import (`await import()`) で biome 静的解析を回避していたが、
noImportCycles の検出対象であり PR #2124 main rebase 後に lint-and-test が fail。

修正: 両 service の共通依存である「期間内 activity_logs のカテゴリ別集計」を
新 module `src/lib/server/services/activity-log-aggregation.ts` に純粋関数として
抽出。これは activity-repo (DB 層) のみに依存し、上位 service には依存しない。

- activity-log-service.getActivityLogs → aggregateActivityLogsByCategory への薄い wrapper
- auto-challenge-service.analyzeWeakCategory → aggregateActivityLogsByCategory を直接呼ぶ
  (dynamic import 撤廃)
- activity-log-service の auto-challenge-service 呼び出しは static import に変更
  (cycle 解消後は dynamic import の必要性なし)

依存グラフ (cycle 完全解消):
  activity-log-service → activity-log-aggregation
  activity-log-service → auto-challenge-service (static)
  auto-challenge-service → activity-log-aggregation
  activity-log-aggregation → activity-repo (DB)

test 追従: auto-challenge-service.test.ts の vi.mock target を
`$lib/server/services/activity-log-service` → `$lib/server/services/activity-log-aggregation`
に変更。テスト assertion 内容は完全維持 (ADR-0006 適合、42 tests pass)。

検証:
- npx biome check src/lib/server/services/ → 0 errors (旧: 2 noImportCycles)
- npx vitest run → 4696/4696 PASS (regression 無し)
- npx svelte-check → 0 errors (既存 13 warnings は本 PR 範囲外)
